+++ include/depthai-shared/common/EepromData.hpp
@@ -1,47 +1,16 @@
 #pragma once
-
-// std
-#include <cstdint>
 #include <vector>
 
 #include "depthai-shared/common/CameraBoardSocket.hpp"
+#include "depthai-shared/common/Point3f.hpp"
+#include "depthai-shared/common/Extrinsics.hpp"
+#include "depthai-shared/common/CameraInfo.hpp"
+#include "depthai-shared/common/StereoRectification.hpp"
 
 // libraries
 #include "nlohmann/json.hpp"
 
 namespace dai {
-
-struct point {
-    float x, y, z;
-    point();
-    point(float _x, float _y, float _z):x(_x),y(_y),z(_z){}
-};
-
-struct Extrinsics {
-    std::vector<std::vector<float>> rotationMatrix;
-    // (x, y, z) pose of destCameraSocket w.r.t currentCameraSocket obtained through calibration
-    point translation;
-    // (x, y, z) pose of destCameraSocket w.r.t currentCameraSocket measured through CAD design
-    point measuredTranslation;
-    CameraBoardSocket toCameraSocket;
-    // FIXME(sachin): Why does fromCameraSocket needs to be vector?
-    std::vector<CameraBoardSocket> fromCameraSockets;
-};
-
-struct StereoRectification {
-    std::vector<std::vector<float>> rectifiedRotationLeft, rectifiedRotationRight;
-    CameraBoardSocket leftCameraSocket, rightCameraSocket;
-    float baseline; // distance between left and right camera in cm
-};
-
-struct CameraInfo {
-    std::vector<std::vector<float>> intrinsicMatrix;
-    std::vector<float> distortionCoeff;
-    Extrinsics extrinsics;
-    double measuredFovRad; //fov in radians
-    // TODO(sachin): Should I add type of distortion model here ?
-};
-
 struct EepromData {
     uint32_t version;
     bool swapLeftRightCam;
@@ -49,6 +18,7 @@ struct EepromData {
     std::unordered_map<CameraBoardSocket, CameraInfo> cameraData;
     StereoRectification stereoRectificationData;
     Extrinsics imuExtrinsics;
+    NLOHMANN_DEFINE_TYPE_INTRUSIVE(EepromData, version, swapLeftRightCam, boardName, boardRev, cameraData, stereoRectificationData, imuExtrinsics);
 };
