+++ include/depthai-shared/common/StereoRectification.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "depthai-shared/common/CameraBoardSocket.hpp"
+#include <vector>
+
+struct StereoRectification {
+    std::vector<std::vector<float>> rectifiedRotationLeft, rectifiedRotationRight;
+    CameraBoardSocket leftCameraSocket, rightCameraSocket;
+    NLOHMANN_DEFINE_TYPE_INTRUSIVE(StereoRectification, rectifiedRotationLeft, rectifiedRotationRight, leftCameraSocket, rightCameraSocket);
+};
